requirements.txt
aiohappyeyeballs 
aiohttp 
aiosignal 
annotated-types 
anyio 
argon2-cffi 
argon2-cffi-bindings 
arrow 
asttokens 
async-lru 
async-timeout 
attrs 
babel 
beautifulsoup4 
bitarray 
bleach 
blinker 
certifi 
cffi 
charset-normalizer 
ckzg 
click 
colorama 
comm 
contourpy 
cryptography 
cycler 
cytoolz 
debugpy 
decorator 
defusedxml 
eth-abi 
eth-account 
eth-hash 
eth-keyfile 
eth-keys 
eth-rlp 
eth-typing 
eth-utils 
exceptiongroup 
executing 
fastjsonschema 
filelock 
flask 
Flask-Login 
flask-sqlalchemy 
fonttools 
fqdn 
frozenlist 
fsspec 
gitdb 
GitPython 
greenlet 
h11 
hexbytes 
httpcore 
httpx 
idna 
importlib-metadata 
importlib-resources 
ipykernel 
ipython 
ipython-pygments-lexers 
isoduration 
itsdangerous 
jedi 
jinja2 
joblib 
json5 
jsonpointer 
jsonschema 
jsonschema-specifications 
# jupyter-client 
# jupyter-core 
# jupyter-events 
# jupyter-lsp 
# jupyter-server 
# jupyter-server-mathjax 
# jupyter-server-terminals 
# jupyterlab 
# jupyterlab-git 
# jupyterlab-pygments 
# jupyterlab-server 
kiwisolver 
MarkupSafe 
matplotlib 
matplotlib-inline 
mistune 
mpmath 
multidict 
narwhals 
nbclient 
nbconvert 
nbdime 
nbformat 
nest-asyncio 
networkx 
notebook-shim 
numpy 
# nvidia-cublas-cu12 
# nvidia-cuda-cupti-cu12 
# nvidia-cuda-nvrtc-cu12 
# nvidia-cuda-runtime-cu12 
# nvidia-cudnn-cu12 
# nvidia-cufft-cu12 
# nvidia-curand-cu12 
# nvidia-cusolver-cu12 
# nvidia-cusparse-cu12 
# nvidia-cusparselt-cu12 
# nvidia-nccl-cu12 
# nvidia-nvjitlink-cu12 
# nvidia-nvtx-cu12 
overrides 
packaging 
pandas 
pandocfilters 
parsimonious 
parso 
pexpect 
pillow 
platformdirs 
plotly 
prometheus-client 
prompt-toolkit 
propcache 
psutil 
ptyprocess 
pure-eval 
py-solc-x 
pycparser 
pycryptodome 
pydantic 
pydantic-core 
pygments 
pyparsing 
python-dateutil 
python-dotenv 
python-json-logger 
pytz 
pyunormalize 
PyYAML 
pyzmq 
referencing 
regex 
requests 
rfc3339-validator 
rfc3986-validator 
rlp 
rpds-py 
# scikit-learn 
scipy 
seaborn 
Send2Trash 
six 
smmap 
sniffio 
soupsieve 
sqlalchemy 
stack-data 
sympy 
terminado 
threadpoolctl 
tinycss2 
tomli 
toolz 
# torch 
tornado 
# traitlets 
# triton 
types-python-dateutil 
types-requests 
typing-extensions 
typing-inspection 
tzdata 
tzlocal 
uri-template 
urllib3 
wcwidth 
web3 
webcolors 
webencodings 
websocket-client 
websockets 
werkzeug 
yarl 
zipp 